backend/requirements.txt
fastapi==0.133.1
uvicorn[standard]==0.41.0
supabase==2.28.0
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
pydantic==2.11.7
email-validator
bcrypt
